Consolidate slugify tests
