+++ tests/test_slugify_conversion.py
@@ -21,3 +21,18 @@ def test_slugified_filename(tmp_path):
     expected_name = slugify(pdf_path.stem) + '.md'
     assert out_file.name == expected_name
     assert out_file.exists()
+
+
+def test_md_filename_ascii(tmp_path):
+    pdf_name = 'ÆÐŐ file.pdf'
+    pdf_path = tmp_path / pdf_name
+    c = canvas.Canvas(str(pdf_path), pagesize=letter)
+    c.drawString(100, 750, 'ascii check')
+    c.save()
+
+    out_dir = tmp_path / 'out_ascii'
+    out_file = convert_pdf_to_md(pdf_path, out_dir, silent=True)
+
+    expected_name = slugify(pdf_path.stem) + '.md'
+    assert out_file.name == expected_name
+    assert all(ord(ch) < 128 for ch in out_file.name), 'Filename contains non-ASCII characters'
